Moved modal shared resize and close to separate files
